backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.34.0
pydantic-settings==2.7.0
prophet==1.2.1
pandas==2.2.0
numpy==1.26.4
redis==5.2.0
httpx==0.28.0
alpaca-py==0.32.0
feedparser==6.0.12
yfinance==0.2.54
sqlmodel==0.0.14
psycopg2-binary==2.9.11
pytest==9.0.2
pyarrow==22.0.0
plotly==6.5.2
lxml==6.0.2
requests==2.32.5
beautifulsoup4==4.14.3
pydantic
scikit-learn
supabase
pytest-cov
coverage
feedparser
praw
fredapi
textblob
email-validator
qstash